working out the issue

+++ src/edu/jsu/mcis/cs310/tas_sp24/dao/DAOUtility.java
@@ -121,7 +121,7 @@ public final class DAOUtility {
         return (int) totalMinutes;
     }
 
-    public static double calculateAbsenteeism(ArrayList<Punch> punchlist, Shift s) {
+    public static BigDecimal calculateAbsenteeism(ArrayList<Punch> punchlist, Shift s) {
 
         ArrayList<Punch> dailyPunches = new ArrayList();
